// src/types/types.ts
export interface Endereco {
  nome: string;
  sobrenome: string;
  rua?: string;
  numero?: string;
  complemento?: string;
  cep?: string;
  planeta: string;
  lote?: string;
  pais?: string;
  cidade?: string;
}

export interface FormValues {
  nome: string;
  sobrenome: string;
  rua?: string;
  numero?: string;
  complemento?: string;
  cep?: string;
  planeta: string;
  lote?: string;
  pais?: string;
  cidade?: string;
}